Assume files with all exec bits set are executable for completions
This checks the st_mode, which we need for other reasons. If it has
the user, group *and* other exec bits set, i.e. a mode of 777 or 755
or similar, we assume it is executable and skip the waccess.

This can have false-positives in some cases:

1. Capabilities
2. noexec filesystems
3. MAC like SELinux?

Of these, capabilities also aren't reflected in access. On a debian
system:

```
sudo capsh --drop=CAP_NET_RAW --shell=/usr/bin/fish -- -c 'test -x /usr/bin/ping && echo is executable; ping 8.8.8.8'
```

prints

> is executable
> exec: Failed to execute process '/usr/bin/ping': No permission. Either suid/sgid is forbidden or you lack capabilities.

and strace confirms:

> faccessat(AT_FDCWD, "/usr/bin/ping", X_OK) = 0

Bash fails similarly, except the error is a more generic "Operation
not permitted"

This leaves noexec filesystems and MAC.

Since this is for completions, it's probably okay to possibly be wrong
in those cases - if it is, the user can just try to execute and will
get an error.

Especially noexec filesystems shouldn't really be in completions
because they'd have to be in $PATH.

I have been unable to test MAC, and from a cursory googling those
would mostly fail the `stat` already.

This speeds up the common case by removing *most* access calls here -
of the 2869 binaries in my $PATH, *1* (one) does not have a mode of
777, 555 or 755. It has 750.

That means completing an empty command will have 2868 fewer access() calls.

This regains all of the speed loss of #9699.