Stack traces: Put "called on" on the same line
When fish prints an error, it will often add a stack trace.

This stack trace could look like

```
test: Missing argument at index 3
-eq 5
      ^
./foo.fish (line 3):
    test $args -eq 5
    ^
in function 'this-errors' with arguments '2'
	called on line 7 of file ./foo.fish
in function 'call-this'
	called on line 1 of file ./foo.fish
in command substitution
	called on line 10 of file ./foo.fish
```

which is a bit long, and the "called on" lines start with tab
characters, which is weird.

So, this shortens it to:

```
test: Missing argument at index 3
-eq 5
      ^
./foo.fish (line 3):
    test $args -eq 5
    ^
in function 'this-errors' with arguments '2' called on line 7 of file ./foo.fish
in function 'call-this' called on line 1 of file ./foo.fish
in command substitution called on line 10 of file ./foo.fish
```

Note: We expose stack traces via `status print-stack-trace`, so this
can break people's scripts if they rely on that.

However, we did a similar change in
7095de628c8f05b2142827a74511cd1c979c5c90,
where we put the function arguments on the same line, and haven't
heard a peep from anyone.