Un-hide documentation of non-fish shell builtins
When writing scripts for other shells, it can be confusing and annoying
that our `man` function shadows other manual pages, for example `exec(1p)`
from [Linux man-pages]. I almost never want to see the fish variant for such
contended cases (which obviuosly don't include fish-specific commands like
`string`, only widely-known shell builtins).

For the contented cases like `exec`, the POSIX documentation is more
substantial and useful, since it describes a (sub)set of languages widely
used for scripting.

Because of this I think we should stop overriding the system's man pages.
Nowadays we offer `exec -h` as intuitive way to show the documentation for
the fish-specific command (note that `help` is not a good replacement because
it uses a web browser).

Looking through the contended commands, it seems like for most of them,
the fish version is not substantially different from the system version.
A notable exception is `read` but I don't think it's a very important one.

So I'm saying: we can sacrifice a bit of the native fish-scripting experience
in exchange for playing nicer with the outside world.  I think the latter
is more important because scripting is not our focus, the way I see it.
So maybe put our manpath at the end.

In lieu of that, we should at least have `exec.rst` reference the system or
POSIX variants, which is what this draft patch does, to set the stage.

[Linux man-pages]: https://www.kernel.org/doc/man-pages/