share/: Use `<?` redirection
This is mostly used instead of `test -r foo && string match bar <foo`
or `set bar (cat foo 2>/dev/null)`.

In doing so it can avoid a weird race condition where the file changes
between the check and the use, and it can avoid an external process or
having to open the file twice.

However it can also cause more work to happen - if you do

```fish
if test -r file
   # do a lot of work with file
end
```

it might be prudent to keep doing that. In the cases here it's usually
just a few `string` calls, which would operate on nothing and so be
pretty quick.