abbr: Allow multiple --command
This means you can define one abbreviation for multiple commands like

```fish
abbr --command={git,hg,svn} co checkout
```

This is surprisingly necessary because any literal match is *also* the
name of the abbr, so you can't actually easily define that otherwise.

You still can't define one replacement with a specific command and
another without without resorting to regexen.

The alternative is to make the command part of the abbr identifier internally, but this implies `--erase` needs it too.