change cd to handle not readable directories: chdir+open instead of open+fchdir
The previous code for the cd builtin used open(.., O_RDONLY) + fchdir to change the working directory.
This fails, if the user has no read permission on the directory (but has the x permission).
Now we first chdir and then try to open the new workin directory. this might fail,
but in this case cd was already successful.

The alternative would be to use O_PATH instead of O_RDONLY, but this is not portable.
As an improvement, we could use O_PATH for supported platforms later.