edit_command_buffer: send indented commandline to editor
One tricky bit is that we need to unindent the resulting command line because
fish already indents. Do this only if the file is properly indented (meaning
at least by the amount fish would use).  Another complication is that we
need to offset cursor positions by the indentation.

This approach exposes "fish_indent --only-indent" and "--only-unindent"
though I don't imagine they are useful for others so I'm not sure if this
is the right place and whether we should even document it.

One alternative is to add "commandline --indented" to handle indentation
transparently.
So  "commandline --indented" would print a indented lines,
and "commandline --indented 'if true' '    echo'" would remove the unecessary
indentation before replacing the commandline.
That would probably simplify the logic for the cursor position offset.